move captive portal code to reportgen module
